requirements.txt
python-dotenv
groq
openai
requests
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
